A script to automate compression and transfer of any folder.
Example:- You have scheduled a daily backup of your Database and
destination to store backup files is a local folder within the same
server. Now you want to automate the transfer of these backup files to a
DR server. In such cases, you can use this script to automate the backup
of these files to a DR server/site.